Refactored SubjectEndpoint into its own class.
